Added a 2nd function (SeqDef2)
This function calculates the synthetic data values, and then does 1 - (this value) in order to give a number that would correspond to how much data is needed (0 being that we have all the data already; and a 1 meaning we have no data, and want as much data as possible on this organism).
Next step is to multiply the syn.imp.vals and syn.dat.vals to get a single number (for each tip) that incorporates these two functions.