Tested both functions further using various values and tree sizes.
I also created an overall value (not within functions.R, just in the testing.R file) that was the product of the synthetic imp.vals and the syn.data.vals... I noticed that this product is heavily influenced by any zero value in either of the function outputs. After this, I took the average of the two values (syn.imp and syn.dat)- in testing.R file.